Integrated Spartan API into SpotifyLogin Sample App
Once authenticated you can now do the following with the app:
- See a list of frank ocean's top songs in the console
- create a new playlist with a name of your choice
