+++ SpartanTestApp/examples/Spartan/SpartanTestApp.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
