+++ SpartanTestApp/examples/Spartan/SpartanTestApp/AppDelegate.swift
