+++ SpartanTestApp/examples/Spartan/SpartanTestApp/Assets.xcassets/AppIcon.appiconset/Contents.json
