+++ SpartanTestApp/examples/Spartan/SpartanTestApp/Base.lproj/LaunchScreen.storyboard
