+++ SpartanTestApp/examples/Spartan/SpartanTestApp/Info.plist
