+++ SpartanTestApp/examples/Spartan/SpartanTestAppTests/Info.plist
